refactor: remove redundant folders
